Add Hiper Theme (Fix Hiperdex (EN), Hipercool (PT-BR), add Hipertoon (PT-BR)) (#17486)
* Merge hipercool and hiperdex to Hiper

* Fix HiperCool auth

* Add hipertoon

* move cookie auth and ratelimit to theme

HiperToon also needs auth cookie

* Move rateLimit out of theme

* Fix versioning

---------
