+++ lib-multisrc/hiper/build.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    alias(kei.plugins.multisrc)
+}
+
+keiyoushi {
+    baseVersionCode = 1
+    libVersion = "1.4"
+}
