+++ src/pt/hipercool/build.gradle.kts
@@ -7,6 +7,7 @@ keiyoushi {
     versionCode = 55
     contentWarning = ContentWarning.NSFW
     libVersion = "1.4"
+    theme = "hiper"
 
     source {
         name = "Hipercool"
