+++ src/pt/hipertoon/build.gradle.kts
@@ -1,0 +1,16 @@
+plugins {
+    alias(kei.plugins.extension)
+}
+
+keiyoushi {
+    name = "Hipertoon"
+    versionCode = 1
+    contentWarning = ContentWarning.MIXED
+    libVersion = "1.4"
+    theme = "hiper"
+
+    source {
+        lang = "pt-BR"
+        baseUrl = "https://hipertoon.com"
+    }
+}
