+++ src/pt/hipertoon/src/eu/kanade/tachiyomi/extension/pt/hipertoon/Hipertoon.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.pt.hipertoon
+
+import eu.kanade.tachiyomi.multisrc.hiper.Hiper
+import keiyoushi.annotation.Source
+
+@Source
+abstract class Hipertoon : Hiper()
